refactor: rename validation
