+++ packages/validation/src/func-result-type.js
