+++ packages/validation/src/is-const.js
