+++ packages/validation/test/fixtures/binop/module.wast
