+++ packages/validation/test/fixtures/br/module.wast
