+++ packages/validation/test/fixtures/result/module.wast
