+++ packages/validation/test/fixtures/store/module.wast
