+++ packages/validation/test/fixtures/unary-operand/module.wast
