+++ packages/validation/test/is-const.js
